Added messaging support, fixed concurrency issue with GLStrings.

+++ src/main/java/us/aaronweiss/pixalia/net/packets/MessagePacket.java
@@ -26,7 +26,7 @@ public class MessagePacket extends Packet {
 	public String hostname() {
 		if (this.packetType.is(PacketType.INBOUND)) {
 			this.ready();
-			return Utils.readString(this.buffer.readInt(), this.buffer);
+			return Utils.readString(this.buffer.readByte(), this.buffer);
 		}
 		return null;
 	}
@@ -34,9 +34,9 @@ public class MessagePacket extends Packet {
 	public String message() {
 		this.ready();
 		if (this.packetType.is(PacketType.INBOUND)) {
-			this.buffer.readBytes(this.buffer.readInt());
+			this.buffer.readBytes(this.buffer.readByte());
 		}
-		return Utils.readString(this.buffer.readInt(), this.buffer);
+		return Utils.readString(this.buffer.readByte(), this.buffer);
 	}
 	
 	public static MessagePacket newInboundPacket(String hostname, String message) {
